iNES/NES2: Extend MapperNum() to support NES2
This means we need 12 bits (16 to include submapper, but we ignore
that for now) to represent the id. Extend all cases to handle the
larger numbers. Add tests to cover the new extneded format handling.

+++ mappers/dummy.go
@@ -10,7 +10,7 @@ type dummyMapper struct {
 	MM     uint8 // mirroring mode - tests can set as needed
 }
 
-func (dm *dummyMapper) ID() uint8 {
+func (dm *dummyMapper) ID() uint16 {
 	return 0
 }
